Y99U82T6TUU52NS2/requirements.txt
plotly
streamlit-aggrid
streamlit-slickgrid
streamlit-quill
streamlit-lexical
streamlit_code_editor
streamlit_calendar
streamlit-echarts
streamlit-pdf-viewer
streamlit-antd-components